Convert data module into package
